Refactor messaging classes and improve logging
Refactored VelocityMessager, GrimBridge, and BungeeMessager classes for better encapsulation and readability. Instead of calling for instance methods, proxy servers and channels are now handled by their respective instances. Logger is also added to provide information when a punishment command is received from the backend server. This update aims to organize code structure and enhance logging for easier debugging and data tracking.

Removed print statements in GrimAPIListener and added a cache system to avoid repeated send of bungee messages. The cache keeps player ID for 10 seconds after messaging, this was put into place to avoid potential spam of commands.

These updates are part of an ongoing effort to improve the efficiency and readability of the codebase.